refactor: move data_warehouses to the new architecture
